Discover Parsing Added And Working
Discover bank and account types added. Considering the work done to make code run with any account type when adding Capital One, this was made very simple. The only changes outside of the new classes, which specify the parsing operations for Discover, is the added menu items to incorporate the new bank type, increased printed information about read files and errors, and restructuring the scraped data validation to take into account values that are a combination of transaction types, denoted with '+' joining combined type names in the true values key.